+++ packages/cli/lib/result-printer.ts
@@ -1,0 +1,5 @@
+import type { Subtitle } from '@subtitles/providers';
+
+export interface ResultPrinter {
+  (subtitle: Subtitle): string;
+}