+++ packages/providers/lib/__tests__/providers.test.js
@@ -1,0 +1,3 @@
+describe('@subtitles/providers', function () {
+    it.todo('needs tests');
+});