+++ packages/providers/lib/providers.ts
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = providers;
-
-function providers() {
-    // TODO
-}