+++ packages/providers/lib/types.ts
@@ -29,9 +29,24 @@ export interface Subtitle {
   language: string;
   source: string; // link to download the subtitle
   media?: Media;
-  file?: ReadableStream;
+  file?: NodeJS.ReadableStream;
   releases?: string[];
   uploader?: string;
   releasedAt?: Date;
   downloads?: number;
+}
+import type { ElementHandle } from 'puppeteer';
+
+export interface SearchFilter {
+  keyword: string;
+  movie?: boolean;
+  tvShow?: boolean;
+}
+
+export interface Provider<T> {
+  search(filter: SearchFilter): Promise<T>;
+}
+
+export interface Scraper<T> {
+  (handle: ElementHandle): Promise<T>;
 }